+++ loop_rate_limiters/__init__.py
@@ -17,7 +17,7 @@
 
 """Loop rate limiters."""
 
-__version__ = "0.4.0"
+__version__ = "0.5.0rc0"
 
 from .async_rate_limiter import AsyncRateLimiter
 from .rate_limiter import RateLimiter